ltc: move external_mem_pool to mpa_desc.c
Moves all initialization of external_mem_pool to mpa_desc.c and also
makes external_mem_pool static as there's no external references.
